fastapi_backend/requirements.txt
# Direct dependencies extracted from pyproject.toml
fastapi[standard]
asyncpg
fastapi-users[sqlalchemy]
pydantic-settings
fastapi-mail
watchdog
geoalchemy2
shapely
pydantic[mypy]
sqlalchemy2-stubs
python-jose

# Development dependencies
pre-commit
ruff
python-dotenv
pytest
pytest-mock
mypy
coveralls
alembic
pytest-asyncio
mkdocs-material
mkdocs-material[imaging]
black
